odb/rcx: mv dbUtil from odb to rcx, its only client
